requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

SQLAlchemy==2.0.29
asyncpg==0.29.0
alembic==1.13.1

pydantic==2.6.4
pydantic-settings==2.2.1
email-validator==2.1.1

python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

python-dotenv==1.0.1
python-multipart==0.0.9

pytest==8.1.1
httpx==0.27.0

pre-commit==4.5.1

